Deprecate nengo.utils.compat
Will be removed completely in the following minor release.
All of the is_* functions were moved to nengo.utils.numpy except
for `is_string` because it's equivalent to `isinstance(var, str)`,
which is already short. Also, all of the other is_* functions exist
in part because of NumPy, so it's actually a pretty good place for
them. We exclude the file from coverage using # pragmas
instead of setup.cfg, so as to avoid needing to remember
to re-modify that file upon final removal of compat.