squash! Deprecate most of nengo.utils.compat
Deprecate all of nengo.utils.compat.

All of the is_* functions were moved to nengo.utils.numpy except
for `is_string` because it's equivalent to `isinstance(var, str)`,
which is already short. Also, all of the other is_* functions exist
in part because of NumPy, so it's actually a pretty good place for
them.